Kalshi by Cemini/requirements.txt
redis
python-dotenv
simplefix
rich
fastapi
uvicorn
httpx
pandas
psycopg2-binary
requests
yfinance
pydantic-settings
feedparser
cryptography
ccxt
sqlalchemy
aiosqlite
tweepy
textblob
python-multipart
